Silence a code scanning warning

+++ kitty/freetype.c
@@ -497,7 +497,7 @@ static inline bool
 render_color_bitmap(Face *self, int glyph_id, ProcessedBitmap *ans, unsigned int cell_width, unsigned int cell_height, unsigned int num_cells, unsigned int baseline) {
     (void)baseline;
     unsigned short best = 0, diff = USHRT_MAX;
-    for (short i = 0; i < self->face->num_fixed_sizes; i++) {
+    for (short i = 0; i < (unsigned short)self->face->num_fixed_sizes; i++) {
         unsigned short w = self->face->available_sizes[i].width;
         unsigned short d = w > (unsigned short)cell_width ? w - (unsigned short)cell_width : (unsigned short)cell_width - w;
         if (d < diff) {
